phase_3/dic_us_accidents/requirements.txt
Django==3.2.16
django-cors-headers==3.10.1
django-cors-middleware==1.5.0
djangorestframework==3.14.0
h5py==3.1.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
scikit-learn==0.24.2
scipy==1.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
pandas~=1.1.5
matplotlib~=3.3.4
seaborn~=0.11.2
protobuf==3.19.6
